clientes/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
grpcio-reflection